noir_debugger::context
Struct DebugContext

pub(crate) struct DebugContext<'a, B: BlackBoxFunctionSolver<FieldElement>> {

    pub(crate) acvm: ACVM<'a, FieldElement, B>,
    current_circuit_id: u32,
    brillig_solver: Option<BrilligSolver<'a, FieldElement, B>>,
    witness_stack: WitnessStack<FieldElement>,
    acvm_stack: Vec<ExecutionFrame<'a, B>>,
    backend: &'a B,
    foreign_call_executor: Box<dyn DebugForeignCallExecutor + 'a>,
    debug_artifact: &'a DebugArtifact,
    breakpoints: HashSet<DebugLocation>,
    source_to_locations: BTreeMap<FileId, Vec<(usize, DebugLocation)>>,
    circuits: &'a [Circuit<FieldElement>],
    unconstrained_functions: &'a [BrilligBytecode<FieldElement>],
    acir_opcode_addresses: AddressMap,
    initial_witness: WitnessMap<FieldElement>,

}

Fields
acvm: ACVM<'a, FieldElement, B>current_circuit_id: u32brillig_solver: Option<BrilligSolver<'a, FieldElement, B>>witness_stack: WitnessStack<FieldElement>acvm_stack: Vec<ExecutionFrame<'a, B>>backend: &'a Bforeign_call_executor: Box<dyn DebugForeignCallExecutor + 'a>debug_artifact: &'a DebugArtifactbreakpoints: HashSet<DebugLocation>source_to_locations: BTreeMap<FileId, Vec<(usize, DebugLocation)>>circuits: &'a [Circuit<FieldElement>]unconstrained_functions: &'a [BrilligBytecode<FieldElement>]acir_opcode_addresses: AddressMapinitial_witness: WitnessMap<FieldElement>
Implementations

impl<'a, B: BlackBoxFunctionSolver<FieldElement>> DebugContext<'a, B>

pub(crate) fn new( blackbox_solver: &'a B, circuits: &'a [Circuit<FieldElement>], debug_artifact: &'a DebugArtifact, initial_witness: WitnessMap<FieldElement>, foreign_call_executor: Box<dyn DebugForeignCallExecutor + 'a>, unconstrained_functions: &'a [BrilligBytecode<FieldElement>], ) -> Self

pub(crate) fn get_opcodes(&self) -> &[Opcode<FieldElement>]

pub(crate) fn get_opcodes_of_circuit( &self, circuit_id: u32, ) -> &[Opcode<FieldElement>]

pub(crate) fn get_witness_map(&self) -> &WitnessMap<FieldElement>

pub(crate) fn overwrite_witness( &mut self, witness: Witness, value: FieldElement, ) -> Option<FieldElement>

pub(crate) fn get_current_debug_location(&self) -> Option<DebugLocation>

pub(crate) fn get_call_stack(&self) -> Vec<DebugLocation>

pub(crate) fn is_source_location_in_debug_module( &self, location: &Location, ) -> bool

pub(crate) fn find_opcode_for_source_location( &self, file_id: &FileId, line: i64, ) -> Option<DebugLocation>

Find an opcode location matching a source code location

pub(crate) fn find_opcode_at_current_file_line( &self, line: i64, ) -> Option<DebugLocation>

pub(crate) fn get_current_source_location(&self) -> Option<Vec<Location>>

Returns the callstack in source code locations for the currently executing opcode. This can be None if the execution finished (and get_current_opcode_location() returns None) or if the opcode is not mapped to a specific source location in the debug artifact (which can happen for certain opcodes inserted synthetically by the compiler). This function also filters source locations that are determined to be in the internal debug module.

fn get_current_file(&self) -> Option<FileId>

Returns the FileId of the file associated with the innermost function on the call stack.

pub(crate) fn get_source_location_for_debug_location( &self, debug_location: &DebugLocation, ) -> Vec<Location>

Returns the (possible) stack of source locations corresponding to the given opcode location. Due to compiler inlining it’s possible for this function to return multiple source locations. An empty vector means that the given opcode location cannot be mapped back to a source location (eg. it may be pure debug instrumentation code or other synthetically produced opcode by the compiler)

pub(crate) fn get_source_call_stack(&self) -> Vec<(DebugLocation, Location)>

Returns the current call stack with expanded source locations. In general, the matching between opcode location and source location is 1 to 1, but due to the compiler inlining functions a single opcode location may expand to multiple source locations.

pub fn debug_location_to_address(&self, location: &DebugLocation) -> usize

Returns the absolute address of the opcode at the given location.

pub fn address_to_debug_location(&self, address: usize) -> Option<DebugLocation>

pub(crate) fn render_opcode_at_location( &self, location: &DebugLocation, ) -> String

fn step_brillig_opcode(&mut self) -> DebugCommandResult

fn handle_foreign_call( &mut self, foreign_call: ForeignCallWaitInfo<FieldElement>, ) -> DebugCommandResult

fn handle_acir_call( &mut self, call_info: AcirCallWaitInfo<FieldElement>, ) -> DebugCommandResult

fn handle_acir_call_finished(&mut self) -> DebugCommandResult

fn handle_acvm_status( &mut self, status: ACVMStatus<FieldElement>, ) -> DebugCommandResult

pub(crate) fn step_into_opcode(&mut self) -> DebugCommandResult

fn get_current_acir_index(&self) -> Option<usize>

fn step_out_of_brillig_opcode(&mut self) -> DebugCommandResult

pub(crate) fn is_executing_brillig(&self) -> bool

pub(crate) fn step_acir_opcode(&mut self) -> DebugCommandResult

pub(crate) fn next_into(&mut self) -> DebugCommandResult

Steps debugging execution until the next source location

pub(crate) fn next_over(&mut self) -> DebugCommandResult

Steps debugging execution until the next source location at the same (or less) call stack depth (eg. don’t dive into function calls)

pub(crate) fn next_out(&mut self) -> DebugCommandResult

Steps debugging execution until the next source location with a smaller call stack depth (eg. returning from the current function)

pub(crate) fn cont(&mut self) -> DebugCommandResult

pub(crate) fn get_brillig_memory(&self) -> Option<&[MemoryValue<FieldElement>]>

pub(crate) fn write_brillig_memory( &mut self, ptr: usize, value: FieldElement, bit_size: BitSize, )

pub(crate) fn get_variables(&self) -> Vec<StackFrame<'_, FieldElement>>

pub(crate) fn current_stack_frame(&self) -> Option<StackFrame<'_, FieldElement>>

fn breakpoint_reached(&self) -> bool

pub(crate) fn is_valid_debug_location(&self, location: &DebugLocation) -> bool

pub(crate) fn is_breakpoint_set(&self, location: &DebugLocation) -> bool

pub(crate) fn add_breakpoint(&mut self, location: DebugLocation) -> bool

pub(crate) fn delete_breakpoint(&mut self, location: &DebugLocation) -> bool

pub(crate) fn clear_breakpoints(&mut self)

pub(crate) fn is_solved(&self) -> bool

pub fn finalize(self) -> WitnessStack<FieldElement>

pub(crate) fn restart(&mut self)