noir_debugger
Module context

Structs

AddressMap

A Noir program is composed by n ACIR circuits |_ m ACIR opcodes |_ Acir call |_ Acir Brillig function invocation |_ p Brillig opcodes

BrilligAddressSpace 🔒

Associates a BrilligFunctionId with the address space. A BrilligFunctionId is found by checking whether an address is between the start_address and end_address

DebugContext 🔒

DebugLocation

DebugProject

DebugStackFrame

ExecutionFrame

RunParams

Enums

DebugCommandResult 🔒

DebugExecutionResult

DebugLocationFromStrError

Functions

add_opcode_locations_map 🔒

build_source_to_opcode_debug_mappings 🔒

Builds a map from FileId to an ordered vector of tuples with line numbers and opcode locations corresponding to those line numbers

initialize_acvm 🔒

is_debug_file_in_debug_crate 🔒